+++ editxt/platform/__init__.py
@@ -33,6 +33,7 @@ def init(platform, use_pdb):
         "font",
         "document",
         "events",
+        "fileref",
         "markdown",
         "pasteboard",
         "text",