+++ editxt/platform/test/fileref.py
@@ -1,0 +1,1 @@
+from ..mac.fileref import FileRef